put latlongtoaddress and vice versa into resource-util
